Bug 1998457 - Part 1: Extend ReciprocalMulConstants for Int64 division. r=spidermonkey-reviewers,iain
Clang and GCC support the non-standard types `__int128_t` and `__uint128_t` on
x64 and arm64 (and other 64-bit architectures, but not for 32-bit architectures).

Alternatively the `[U]Int128` classes from "js/src/builtin/temporal/Int128.h" can
be used.

Differential Revision: https://phabricator.services.mozilla.com/D271436

UltraBlame original commit: cca4dd2b12477ea03dba85b4420f150d1745f46e

+++ js/src/jit/MacroAssembler.cpp
@@ -2062,7 +2062,8 @@ void MacroAssembler::loadInt32ToStringWithBase(
       
       
 
-      auto rmc = ReciprocalMulConstants::computeUnsignedDivisionConstants(base);
+      auto rmc = ReciprocalMulConstants::computeUnsignedDivisionConstants(
+          uint32_t(base));
 
       
       mulHighUnsigned32(Imm32(rmc.multiplier), input, scratch1);
